+++ Blurred-ios/Blurred-ios-Bridging-Header.h
@@ -6,5 +6,4 @@
 //  Copyright © 2020 BlurrMC. All rights reserved.
 //
 
-#import "GradientNode.h"