+++ CHANGELOG.md
@@ -111,6 +111,7 @@
     * Added video group
     * Added bridging header (to support gradient node objective c files)
     * Cleaned up code
+    * Added showMessage function instead of having whole new function for every different alert message.
 ## v0.3 (PRERELEASE)
 * Changed scale factor of avatar image in channel view.
 * Added description field